重写配置文件，添加对 floodgate 的接管支持

+++ core/src/main/java/moe/caa/multilogin/core/ohc/RetryInterceptor.java
@@ -14,9 +14,9 @@ import java.util.concurrent.TimeUnit;
  */
 public class RetryInterceptor implements Interceptor {
     private final int retry;
-    private final int delay;
+    private final long delay;
 
-    public RetryInterceptor(int retry, int delay) {
+    public RetryInterceptor(int retry, long delay) {
         this.retry = retry;
         this.delay = delay;
     }
